// src/app/model/worker.ts
export class Worker {
  firstname: string;
  lastname: string;
  login: string;
  password: string;
  email: string;
  responsibilities: string;
  position: string;
  dateofemployment: Date;
  dateofbirth: Date;
  salary: number;
  activity: string;
  pesel: number;
  workinghours: string;
  idWorker: number;
  privileges : number;
}